refactor(metering): move realtime usage under observability; extract generic ticker
src/realtime-usage.ts → split into:
  src/observability/ticker.ts   — generic startTicker<T> (any streaming surface)
  src/observability/realtime.ts — voice + phone domain metering (uses ticker)

ticker.ts is surface-agnostic: callers supply an onTick(durationMs, bucketStartMs)
callback; the interval and idempotent stop() are handled generically. A future
screen-capture or API-gateway meter can reuse it without touching realtime.ts.

tests/realtime-usage.test.ts → tests/observability/realtime.test.ts (mirrors src layout).
Import paths updated in voice-agent.ts and conversation-server.ts.
